chane size

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -8,7 +8,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t int i;
 
 	if (array != NULL && size != NULL && action != NULL)
 	{
